requirements.txt
torch==1.8.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.8.0
datasets==1.13.3
transformers==4.11.3
librosa
jiwer
tensorboard==2.7.0
hydra-core==1.1.1
